Solution to Convert Decimal to Binary

+++ Convert Decimal to Binary@@67882a731b4392d8fc1577f5/code.c
@@ -7,7 +7,7 @@ int main() {
     int b=0;
     int d=0;
     while(a>0){
-        d=a%10;
+        d=(int)a%10;
         if(d%2==0){
             b=b*10;
         }
